requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
matplotlib==3.7.2
einops==0.6.1
timm==0.9.2
h5py==3.9.0
pyyaml==6.0.1
wandb==0.15.8
tqdm==4.65.0
mne==1.4.2
pandas==2.0.3
